Experimental support for detecting bonded pet ghosts.
Dead bonded pets appear as "ghosts" in the world, but, like many other
things in UO, regular player ghosts are not a *flag* but a *model*.
It appears that rather than add an extra model for every possible
tameable, OSI elected to add a flag, strictly for bonded pets.

Until now, RE has not had supported for detecting if a pet is dead - a
dead pet just looks like any other mobile to RE.

This commit is an attempt to process that flag, which is supplied by an
overload of the 0x19 (originally stat locks) subcommand of the 0xBF
(general information) packet.

To use, simply check the boolean Mobile.IsDeadPet property in your
script.

This has been lightly tested on ServUO and nowhere else.

UOSteam script support is included via a new "isdeadpet" command, but
has not been tested.

This support is supplied via new flags/commands rather than the regular
"dead" properties for fear of introducing odd backward compatibility
problems with scripts that just assume pet mobiles can't be "dead".